Small corrections
